requirements.txt
# requirements.txt
google-cloud-aiplatform
requests
pandas
pytrends
python-dotenv
fastapi
uvicorn[standard]
google-cloud-firestore